Implement new search method
scryfall has an autocomplete endpoint that returns a list of 20 card names best matching our query.  Use this api then fetch the 20 cards in the list.

This removes the need for paging library as old data was any card that contained the letters in order of the query which could produce 100's of results.

Also, will help us to store the results in db for subsequent searches since this separates search logic and fetching of specific cards.